Refactor: break topology into more modules for readability.
